Change aggregate member object names to include the aggregate nesting
Modify the construction of the OpenMCT name for aggregate members to
include the path from the aggregate parameter to the member. So "x"
inside "position" inside "roverPose" gets an OpenMCT name of
"rosePose_position_x". Add a special XTCE alias namespace to cause
a parameter to be omitted from the OpenMCT object tree. Clean up
iteration over parameter data to use iteration functions instead of
for loops.